optimize bounds and face edge node construction
